wip: file restructure
