+++ lib/cli/dev.js
@@ -1,6 +1,6 @@
 import { serve } from './serve.js';
 import { watch } from './watch.js';
-import { writePageFiles } from '../files.js';
+import { writePageFiles } from '../core/files.js';
 
 export function dev(config, _cayo) {
   // for (const [,page] of _cayo.pages) {
