+++ lib/cli/index.js
@@ -8,15 +8,15 @@ import {
   writeComponentFile,
   cleanCayoPath,
   writeTemplateCSS,
-} from '../files.js';
+} from '../core/files.js';
 
 import { 
   hash,
-} from '../utils.js';
+} from '../core/utils.js';
 
-import { loadConfig, checkConfigPaths } from '../config.js';
-import * as compile from '../compile/index.js';
-import { logger } from '../logger.js';
+import { loadConfig, checkConfigPaths } from '../core/config.js';
+import * as compile from '../core/compile/index.js';
+import { logger } from '../core/logger.js';
 
 import { dev } from './dev.js';
 import { build } from './build.js';
