+++ lib/cli/watch.js
@@ -1,14 +1,14 @@
 import path from 'path';
 import chokidar from 'chokidar';
 import chalk from 'chalk';
-import { logger } from '../logger.js';
-import * as compile from '../compile/index.js';
+import { logger } from '../core/logger.js';
+import * as compile from '../core/compile/index.js';
 import { 
   writePageFiles,
   writeComponentFile,
   cleanCayoPath,
   writeTemplateCSS,
-} from '../files.js';
+} from '../core/files.js';
 
 export function watch(config, _cayo) {
