+++ lib/core/codegen.js
