+++ lib/core/compile/cayoComponent.js
@@ -8,7 +8,7 @@ import { fileURLToPath } from 'url';
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 export async function compileCayoComponent(config) {
-  const source = await fs.readFile(path.resolve(__dirname, '../components/cayo.svelte'), 'utf8');
+  const source = await fs.readFile(path.resolve(__dirname, '../../components/cayo.svelte'), 'utf8');
   let filename = 'cayo.svelte';
   const component = await preprocess(source, sveltePreprocess(), { filename });
   const {
