+++ lib/core/compile/components.js
