+++ lib/core/compile/dependencies.js
