+++ lib/core/compile/index.js
