+++ lib/core/compile/layout.js
