+++ lib/core/compile/pages.js
