+++ lib/core/component.js
