+++ lib/core/config.js
