+++ lib/core/logger.js
