+++ lib/core/page.js
@@ -1,5 +1,5 @@
 import { Component } from "./component.js";
-import { prerender } from "./prerender.js";
+import { prerender } from "./render/prerender.js";
 import { default as fse } from 'fs-extra';
 import { build } from './rollup.js'
 import { hash } from './utils.js';
