+++ lib/core/render/prerender.js
@@ -1,45 +1,45 @@
 import fs from 'fs-extra';
 import path from 'path';
 import { JSDOM } from 'jsdom';
-import { Renderer } from './renderer.js';
-import { generateCayoRuntime } from './codegen.js';
 import chalk from 'chalk';
+import { Renderer } from './renderer.js';
+import { logger } from '../logger.js';
+import { generateCayoRuntime } from '../codegen.js';
 
-export function prerender(Template, pages, componentModules, config, logger) {
-  const template = Template.render();
-  const renderer = new Renderer(template);
+export async function prerender(page, cayoComponents, config) {
+  const renderer = new Renderer(page.layout);
   const componentList = new Set();
 
-  // const prerendered = {};
-  // Render page, parse html, and save its deps
-  const prerendered = Object.entries(pages).reduce(
-    (prerendered, [pathname, page]) => {
-      // Render page
-      const content = renderer.render(page, config);
-      // Postprocess the content, get deps and inject dep references
-      const { html, css, js, entry, components } = processPage(content, page, Object.keys(componentModules), config, logger);
-      prerendered[pathname] = {
-        html,
-        css, 
-        js,
-        entry,
-        components,
-        ...page,
-      }
-      Object.keys(components).forEach(component => componentList.add(component))
-      return prerendered;
-    }, {}
-  );
+  
 
-  return { 
-    prerendered,
-    componentList,
-    template,
+  const content = await renderer.render(page, config);
+  // Postprocess the content, get deps and inject dep references
+  const { html, css, js, entry, components } = processPage(
+    content, 
+    page, 
+    cayoComponents, 
+    config, 
+    logger
+  );
+  const prerender = {
+    html,
+    css, 
+    js,
+    entry,
+    components,
+  }
+
+  Object.keys(components).forEach(component => componentList.add(component))
+
+  return {
+    ...prerender,
+    components: [...componentList],
   }
 }
 
 // Derive JS dependencies from the prerendered html
-export function processPage(content, page, componentNames, config, logger) {
+export function processPage(content, page, cayoComponents, config, logger) {
+  const componentNames = Object.keys(cayoComponents);
   const tags = findDocumentTags(content.html);
   const dom = new JSDOM(content.html);
   const { document } = dom.window;
@@ -50,9 +50,9 @@ export function processPage(content, page, componentNames, config, logger) {
     if (el.dataset.cayoId !== '') {
       cayoIds.push(el.dataset.cayoId);
     } else {
-      logger.info(
-        chalk.red(`Cayo component instance without a name found`) + chalk.dim(` ${page.filePath}`), 
-        { timestamp: true, clear: true, }
+      logger.error.warn(
+        chalk.red(`Cayo component instance without a name found`) + chalk.dim(` ${page.name}`), 
+        { timestamp: true, clear: false, }
       );
     }
   });
@@ -63,11 +63,11 @@ export function processPage(content, page, componentNames, config, logger) {
   const components = cayoIds.reduce((components, id) => {
     let name = id.match(componentNameRegex).groups.name;
     if (!componentNames.includes(name)) {
-      logger.warn(
+      logger.error.warn(
         chalk.red(
           `Cayo component with name '${name}' does not exist but is trying to be rendered`
-        ) + chalk.dim(` ${page.filePath}`), 
-        { timestamp: true, clear: true, }
+        ) + chalk.dim(` ${page.name}`), 
+        { timestamp: true, clear: false, }
       );
       return components;
     }
@@ -104,7 +104,7 @@ export function processPage(content, page, componentNames, config, logger) {
     if (entryScript) {
       entryScript.remove();
     } else {
-      logger.info(
+      logger.log.info(
         chalk.bgRed.white(`No entry placeholder in template file.`) + chalk.dim(` Cayo components will not render.`), 
         { timestamp: true }
       );
@@ -114,11 +114,11 @@ export function processPage(content, page, componentNames, config, logger) {
   if (userEntryFile[0] === '/') {
     userEntryFile = userEntryFile.substring(1);
   } else if (userEntryFile && userEntryFile[0] !== '/') {
-    logger.warn(
+    logger.error.warn(
       chalk.red(
         `Entry file path '${userEntryFile}' requires a leading slash and to be relative to src`
-      ) + chalk.dim(` ${page.filePath}`), 
-      { timestamp: true, clear: true, }
+      ) + chalk.dim(` ${page.name}`), 
+      { timestamp: true, clear: false, }
     );
   }
 
@@ -129,7 +129,7 @@ export function processPage(content, page, componentNames, config, logger) {
     );
   
     if (!fs.pathExistsSync(entryFilePath)) {
-      console.error(`Can't read entry file ${userEntryFile} in ${page.modulePath}`);
+      console.error(`Can't read entry file ${userEntryFile} in ${page.name}`);
     } else {
       entry.path = entryFilePath;
     }
