+++ lib/core/render/renderer.js
