+++ lib/core/utils.js
