+++ lib/core/vite.config.js
